requirements/base.txt
# Base Project Requirements

dj-static == 0.0.6
Django == 1.7
django-configurations == 0.8
django-dotenv == 1.2
django-grappelli == 2.6.1
django-heroku-postgresify == 0.3
django-model-utils == 2.0.3
djangorestframework == 2.4.3
Pillow == 2.4.0
psycopg2 == 2.5.2


